requirements.txt
# Python3.12 project
Jinja2==3.1.3
python-http-client==3.3.7
requests==2.31.0
sendgrid==6.11.0
tabulate==0.9.0
pyyaml==6.0.1
